string map: stop copying data
The primary use case is to work with owned-strings read from the wire,
so already allocated somewhere.

This means if we give a string coming from a literal string it will
segfault.

There's maybe a way to check for that type of string and avoid the call
to free in that case.