iterator: take an allocator in `scan()`
A single query can return a lot of data so having a single arena for the
entire iterator could lead to a lot of memory usage.

Instead pass an allocator for every scan call; the caller is then
responsible for deallocating the data.

This can be done trivially with a ArenaAllocator