client: improve cprepare and execute
Take a struct/tuple in cprepare which will be used to compute a list of
OptionID for the bound values.

This metadata will be stored in the client and later call to execute
will then compare its `args` with the metadata stored: if incompatible
the call fails.

This forces the caller to make sure the same types are used for a single
prepared query id.

Unfortunately this means that we can't execute a statement prepared by
another client (remember that the prepared statement are the same for
the entire node). Maybe this can be solved by having another variant of
execute which is unsafe.