client: don't validate the type of arrays/slices
computeValues computes a list of OptionID which are checked against the
prepared metadata column specs in client.execute.

However with arrays and slices (aside from the special cases [16]u8 and
[]const u8) it's not possible to know if it's a set or a list in the
Cassandra table.  This is actually not a problem this the only reason we
compute these options is for checking and we don't actually send them to
cassandra.

So instead of trying to know if it's a set or list now we just ignore
the option when processing an array or a slice.