client: test using a real cassandra node if requested
Add the capability to test the client with a real cassandra node.

This can be tested with `zig build test -Dwith_cassandra`, without the
option the tests are skipped.

This gating is necessary because running tests against a real cassandra
node is not always possible or desirable.

The file `casstest.zig` provides helpers types and functions to create a
test client; these helpers are also used in `example.zig`.